email-service/app/notifications/serializers.py
from rest_framework import serializers
from .models import ContactMessage, NotificationLog

class ContactMessageSerializer(serializers.ModelSerializer):
    class Meta:
        model = ContactMessage
        fields = '__all__'

class NotificationLogSerializer(serializers.ModelSerializer):
    class Meta:
        model = NotificationLog
        fields = '__all__'
